# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Neon Murer CMS API
  description: |
    Vollständige REST API für das Neon Murer Content Management System.
    
    ## Authentifizierung
    Die meisten Endpoints erfordern JWT-Token Authentication über den `Authorization: Bearer <token>` Header.
    
    ## Fehlerbehandlung
    Alle Endpoints verwenden einheitliche HTTP-Status-Codes und JSON-Antworten.
    
    ## Rate Limiting
    API-Aufrufe sind auf 100 Requests pro Minute pro IP-Adresse begrenzt.
    
  version: 1.0.0
  contact:
    name: Neon Murer AG
    url: https://neonmurer.ch
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001/api
    description: Development Server
  - url: https://cms.neonmurer.ch/api  
    description: Production Server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: "Fehlermeldung"
        details:
          type: object
...
    Customer:
      type: object
      properties:
        id:
          type: string
          example: "customer_123"
        company:
          type: string
          example: "Mustermann GmbH"
        firstName:
          type: string
          example: "Max"
        lastName:
          type: string
          example: "Mustermann"
        email:
          type: string
          format: email
          example: "max@mustermann.ch"
        phone:
          type: string
          example: "+41 44 123 45 67"
        address:
          type: string
          example: "Musterstrasse 123"
        city:
          type: string
          example: "Zürich"
        zipCode:
          type: string
          example: "8001"
        country:
          type: string
          example: "Schweiz"
        status:
          type: string
          enum: [ACTIVE, INACTIVE, PROSPECT]
          example: "ACTIVE"
        contacts:
          type: array
          items:
            $ref: '#/components/schemas/Contact'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Contact:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        position:
          type: string
        department:
          type: string
        isActive:
          type: boolean
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: Alle Kunden abrufen
      security:
        - bearerAuth: []
      parameters:
        - name: includeContacts
          in: query
          schema:
            type: boolean
            default: false
        - name: includeActivities
          in: query
          schema:
            type: boolean
            default: false
        - name: status
          in: query
          schema:
            type: string
            enum: [ACTIVE, INACTIVE, PROSPECT]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Kunden erfolgreich abgerufen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  customers:
                    type: array
                    items:
                      $ref: '#/components/schemas/Customer'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
                    
    post:
      tags: [Customers]
      summary: Neuen Kunden erstellen
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                company:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                address:
                  type: string
                city:
                  type: string
                zipCode:
                  type: string
                country:
                  type: string
                  default: "Schweiz"
                status:
                  type: string
                  enum: [ACTIVE, INACTIVE, PROSPECT]
                  default: PROSPECT
      responses:
        '201':
          description: Kunde erfolgreich erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  customer:
                    $ref: '#/components/schemas/Customer'
...
  /customers/stats:
    get:
      tags: [Customers]
      summary: Kunden-Statistiken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Statistiken erfolgreich abgerufen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  stats:
                    type: object
                    properties:
                      total:
                        type: integer
                      active:
                        type: integer
                      inactive:
                        type: integer
                      prospects:
                        type: integer
  /customers/{id}:
    get:
      tags: [Customers]
      summary: Einzelnen Kunden abrufen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Kunde erfolgreich abgerufen
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  customer:
                    $ref: '#/components/schemas/Customer'
        '404':
          description: Kunde nicht gefunden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
                
    put:
      tags: [Customers]
      summary: Kunden aktualisieren
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                company:
                  type: string
                firstName:
                  type: string
                lastName:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                address:
                  type: string
                city:
                  type: string
                zipCode:
                  type: string
                country:
                  type: string
                status:
                  type: string
                  enum: [ACTIVE, INACTIVE, PROSPECT]
      responses:
        '200':
          description: Kunde erfolgreich aktualisiert
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  customer:
                    $ref: '#/components/schemas/Customer'
                    
    delete:
      tags: [Customers]
      summary: Kunden löschen
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Kunde erfolgreich gelöscht
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Kunde erfolgreich gelöscht"
                    
  # MEDIA ENDPOINTS
...